Remove incorrect comment from server help package output.
When running server help package the following is output at the end:

```
 at 72 chars -- leading tab-----------
```

this comment goes in the nls file to ensure that the width of options
in the help output is no more than 72 characters. However the proceeding
line was incorrectly terminated with \n\ which say print new line and then
turn off the next new line character. This means the # that indicates the
start of a comment is treated as part of the prior key and results in a
comment appearing in the output. This has been in Liberty for at least 5
years.