Disable OpenAPI UI FAT screen recordings
Currently our OpenAPI FAT tests are set up to record a screen capture of
the UI as the test runs, and to save it if any tests fail.

However, we're sometimes seeing failures where the Vnc container which
does the recording doesn't start correctly. I'm not sure if that's the
only problem or whether the Vnc container error is masking another
failure.

To get more information on these failures, I'd like to disable screen
recording for now so that we can see if there's an actual failure
underneath.

It's possible that in future we'll need to reenable screen recording if
we need it for debugging.