Remove default value from verify hostname description.
The description of verifyHostname includes a statement of the default.
This is unnecessary because our documentation and xml schema show the
default value, and also wrong because the default is true, but the
description says it is false. Rather than update the description to
have the correct default it is better to remove it and rely on the
generated documentations default value cell to cover the default value.

Fix #30308