MP OpenAPI: Warn if MP Config properties ignored
Warn if MP Config properties for module inclusion and exclusion or for
Info config are ignored because configuration is provided in the
server.xml file.

Use an appropriate message depending on whether the configuration in MP
Config is equivalent to or conflicts with the configuration in the
server.xml file.

Previously we only read the Info from MP Config when we needed to. Now
we need to check if we need to warn about MP Config being ignored after
we successfully read info config from server.xml.

Adding these warnings for the Info config requires refactoring the MP
Config reading code, so also move it into OpenAPIInfoConfig so that all
the Info reading code is in one place.