Add mpHealth-4.0 FAT TCK for JAVA EE repeats named io.openliberty.microprofile.health.4.0.internal.javax_fat_tck
Under the covers this pulls the MP Health 3.1 TCK from maven. This works as mpHealth-4.0 and mpHealth-3.1 are exactly the same besides the Javax and Jakarta difference. This allows us to use the 3.1 TCK to run against mpHealth-4.0 under a Java EE environment.
Any future mpHealth version that also requires backwards support may not be able to leverage this unless Javaxified version of the TCK is published (either through the project or internally within IBM).